pyproject.toml
[tool.poetry]
name = "auto-fcmobile"
version = "0.1.0"
description = ""
authors = ["felixcapuano <felix.capuano@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pywinauto = "^0.6.8"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
start = "auto_fcmobile.main:main"